pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qssh"
version = "0.2.3"
description = "Quick SSH session manager - save your VM credentials and connect with a single command"
readme = "README.md"
license = "MIT"
authors = [
    {name = "joan-code6"}
]
keywords = ["ssh", "session", "manager", "vm", "cli", "terminal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking",
    "Topic :: Utilities",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "paramiko>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/joan-code6/qssh"
Repository = "https://github.com/joan-code6/qssh"
Issues = "https://github.com/joan-code6/qssh/issues"

[project.scripts]
qssh = "qssh.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
qssh = ["py.typed"]
